Split codebase up and got macros in macros partially working
Macro expansions now scope their end parenthesis, but arguments are
still parsed the old way, so the form #(MACRO 0, #(VALUE)) works, but
not the form #(MACRO 0, #(OTHER_MACRO 0, 1)) doesn't.
Also made it so empty lines are trimmed-out, but I need to implement a
flag to toggle, and more globally revamp the flag module.